# [core-rest-pipeline] Cache pipeline function