--- a/sdk/core/perf-tests/core-rest-pipeline/test/index.spec.ts
+++ b/sdk/core/perf-tests/core-rest-pipeline/test/index.spec.ts
@@ -1,8 +1,11 @@
 import { PerfProgram, selectPerfTest } from "@azure/test-utils-perf";
 import { BearerTokenAuthenticationPolicyChallengeTest } from "./bearerTokenChallengeAuthenticationPolicy/wwwChallenge.spec";
+import { PipelineThroughputTest } from "./throughput.spec";
 
 console.log("=== Starting the perf test ===");
 
-const perfProgram = new PerfProgram(selectPerfTest([BearerTokenAuthenticationPolicyChallengeTest]));
+const perfProgram = new PerfProgram(
+  selectPerfTest([BearerTokenAuthenticationPolicyChallengeTest, PipelineThroughputTest])
+);
 
 perfProgram.run();