### Packages impacted by this PR

`@azure/core-rest-pipeline`

### Issues associated with this PR

N/A

### Describe the problem that is addressed by this PR

The pipeline goes through the process of reducing its ordered policies into a single `SendRequest` function each time the `sendRequest` method is called. This accounts for more than 5% of the `sendRequest` method's execution time.

This PR adds a `WeakMap` cache to weakly associate the resulting `SendRequest` function as a hidden property of the HTTP client. The `sendRequest` function will first search the cache and use the associated function if one exists.

When policies are added/removed to the pipeline, a new `WeakMap` instance is created, invalidating all previously-cached entries.

**WITH** optimization:

```
=== test mode, iteration 1. Logs every 1s ===
ElapsedTime     Current         Total           Average
=== test mode, results of iteration 1 ===
Completed 2,590,107 operations in a weighted-average of 10.00s (259,010.64 ops/s, 0.000 s/op)
```

**WITHOUT** optimization:
```
=== test mode, iteration 1. Logs every 1s ===
ElapsedTime     Current         Total           Average
=== test mode, results of iteration 1 ===
Completed 2,455,799 operations in a weighted-average of 10.00s (245,579.85 ops/s, 0.000 s/op)
```

An increase of ~13.5k op/s is around a 5.4% improvement.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?

This is the first thing that came to mind.

### Are there test cases added in this PR? _(If not, why?)_

Yes, I added a new perf test to measure raw pipeline throughput. The test does not use any asynchronous IO operations and resolves responses immediately to test the performance of the pieline alone. It sends the same exact request and response shapes every time so that the `sendRequest` method will be highly monomorphisable by the engine.

### Provide a list of related PRs _(if any)_

N/A

### Command used to generate this PR:**_(Applicable only to SDK release request PRs)_

N/A

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so, create an Issue in the [Autorest/typescript](https://github.com/Azure/autorest.typescript) repository and link it here)_
- [x] Added a changelog (if necessary)